tabl2

+++ prog.py
@@ -130,6 +130,7 @@ class Dictionary:
         else:
             self.message['text'] = 'введите слово и значение'
         self.get_words()
+    #удаление
     def delete_word(self):
         self.message['text'] = ''
         try:
@@ -139,7 +140,11 @@ class Dictionary:
             return
         self.message['text'] = ''
         typeofmaterial= self.tree.item(self.tree.selection())['text']
-        query = f'DELETE FROM typesofmaterials WHERE id = {typeofmaterial}'
+        global a
+        if a==1:
+            query = f'DELETE FROM typesofmaterials WHERE id = {typeofmaterial}'
+        if a==2:
+            query = f'DELETE FROM materials WHERE id = {typeofmaterial}'
         self.run_query(query,False, )
         self.message['text'] = 'Слово {} успешно удалено'.format(typeofmaterial)
         self.get_words()
@@ -151,30 +156,78 @@ class Dictionary:
         except IndexError as e:
             self.message['text'] = 'Выберите строку для изменения'
             return
-        word = self.tree.item(self.tree.selection())['text']
-        old_meaning = self.tree.item(self.tree.selection())['values'][0]
-        self.edit_wind = Toplevel()
-        self.edit_wind.title = 'Изменить значение'
-
-        Label(self.edit_wind, text = 'Прежнее id:').grid(row = 0, column = 1)
-        Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = word), state = 'readonly').grid(row = 0, column = 2)
         
-        Label(self.edit_wind, text = 'Новое id:').grid(row = 1, column = 1)
-        # предзаполнение поля
-        new_word = Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = word))
-        new_word.grid(row = 1, column = 2)
+        global a
+        if a==1:
+            word = self.tree.item(self.tree.selection())['text']
+            old_meaning = self.tree.item(self.tree.selection())['values'][0]
+            self.edit_wind = Toplevel()
+            self.edit_wind.title = 'Изменить значение'
+            Label(self.edit_wind, text = 'Прежнее id:').grid(row = 0, column = 1)
+            Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = word), state = 'readonly').grid(row = 0, column = 2)
+        
+            Label(self.edit_wind, text = 'Новое id:').grid(row = 1, column = 1)
+            # предзаполнение поля
+            new_word = Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = word))
+            new_word.grid(row = 1, column = 2)
 
 
-        Label(self.edit_wind, text = 'Прежнее вид матерьяла:').grid(row = 2, column = 1)
-        Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = old_meaning), state = 'readonly').grid(row = 2, column = 2)
+            Label(self.edit_wind, text = 'Прежнее вид матерьяла:').grid(row = 2, column = 1)
+            Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = old_meaning), state = 'readonly').grid(row = 2, column = 2)
  
-        Label(self.edit_wind, text = 'Новое вид матерьяла:').grid(row = 3, column = 1)
-        # предзаполнение поля
-        new_meaning= Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = old_meaning))
-        new_meaning.grid(row = 3, column = 2)
+            Label(self.edit_wind, text = 'Новое вид матерьяла:').grid(row = 3, column = 1)
+            # предзаполнение поля
+            new_meaning= Entry(self.edit_wind, textvariable = StringVar(self.edit_wind, value = old_meaning))
+            new_meaning.grid(row = 3, column = 2)
 
-        Button(self.edit_wind, text = 'Изменить', command = lambda: self.edit_records(new_word.get(), word, new_meaning.get(), old_meaning)).grid(row = 4, column = 2, sticky = W)
-        self.edit_wind.mainloop()
+            Button(self.edit_wind, text = 'Изменить', command = lambda: self.edit_records(new_word.get(), word, new_meaning.get(), old_meaning)).grid(row = 4, column = 2, sticky = W)
+            self.edit_wind.mainloop()
+        if a==2:
+            # Прежние значения
+            selected_item = self.tree.selection()[0]  # Предполагается, что выбран один элемент
+            word = self.tree.item(selected_item)['text']
+            old_values = self.tree.item(selected_item)['values']
+        
+            self.edit_wind = Toplevel()
+            self.edit_wind.title('Изменить значения')
+            Label(self.edit_wind, text='Прежнее id:').grid(row=0, column=1)
+            Entry(self.edit_wind, textvariable=StringVar(self.edit_wind, value=word), state='readonly').grid(row=0, column=2)
+        
+            Label(self.edit_wind, text='Прежний размер:').grid(row=1, column=1)
+            Entry(self.edit_wind, textvariable=StringVar(self.edit_wind, value=old_values[0]), state='readonly').grid(row=1, column=2)
+        
+            Label(self.edit_wind, text='Прежняя характеристика:').grid(row=2, column=1)
+            Entry(self.edit_wind, textvariable=StringVar(self.edit_wind, value=old_values[1]), state='readonly').grid(row=2, column=2)
+        
+            Label(self.edit_wind, text='Прежнее количество:').grid(row=3, column=1)
+            Entry(self.edit_wind, textvariable=StringVar(self.edit_wind, value=old_values[2]), state='readonly').grid(row=3, column=2)
+        
+            Label(self.edit_wind, text='Прежний вид матерьяла:').grid(row=4, column=1)
+            Entry(self.edit_wind, textvariable=StringVar(self.edit_wind, value=old_values[3]), state='readonly').grid(row=4, column=2)
+
+            # Новые значения
+            Label(self.edit_wind, text='Новое id:').grid(row=0, column=3)
+            new_id = Entry(self.edit_wind)
+            new_id.grid(row=0, column=4)
+
+            Label(self.edit_wind, text='Новый размер:').grid(row=1, column=3)
+            new_size = Entry(self.edit_wind)
+            new_size.grid(row=1, column=4)
+
+            Label(self.edit_wind, text='Новая характеристика:').grid(row=2, column=3)
+            new_characteristic = Entry(self.edit_wind)
+            new_characteristic.grid(row=2, column=4)
+
+            Label(self.edit_wind, text='Новое количество:').grid(row=3, column=3)
+            new_quantity = Entry(self.edit_wind)
+            new_quantity.grid(row=3, column=4)
+
+            Label(self.edit_wind, text='Новый вид матерьяла:').grid(row=4, column=3)
+            new_material_type = Entry(self.edit_wind)
+            new_material_type.grid(row=4, column=4)
+
+            Button(self.edit_wind, text='Изменить', command=lambda: self.edit_records1(new_id.get(), word, new_size.get(), old_values[0], new_characteristic.get(), old_values[1], new_quantity.get(), old_values[2], new_material_type.get(), old_values[3])).grid(row=5, column=2, sticky=W)
+            self.edit_wind.mainloop()
     #новое
     def novoe(self):
         self.message['text'] = ''
@@ -215,8 +268,8 @@ class Dictionary:
             ttk.Button(frame, text = 'Сохранить', command = self.add_word).grid(row = 6, columnspan = 2, sticky = W + E)
     # внесение изменений в базу
     def edit_records(self, new_word, word, new_meaning, old_meaning):
+      
         query = f"""UPDATE typesofmaterials SET id = {new_word}, typeofmaterial = '{new_meaning}'WHERE id = {word} AND typeofmaterial = '{old_meaning}'"""
-        
         self.run_query(query,False)
         self.edit_wind.destroy()
         self.message['text'] = 'слово {} успешно изменено'.format(word)
@@ -226,15 +279,30 @@ class Dictionary:
         for element in records:
             self.tree.delete(element)
         x=self.poiskk.get()
+        global a 
         if x.isdigit():
-            query=f"""SELECT * FROM typesofmaterials WHERE id = {self.poiskk.get()}  """
+            if a==1:
+                query=f"""SELECT * FROM typesofmaterials WHERE id = {self.poiskk.get()}  """
+            if a==2:
+                query=f"""SELECT * FROM materials WHERE id = {self.poiskk.get()} or quantity = {self.poiskk.get()} or material_id ={self.poiskk.get()}"""
         else:
-            query=f"""SELECT * FROM typesofmaterials WHERE  typeofmaterial = '{self.poiskk.get()}' """
+            if a==1:
+                query=f"""SELECT * FROM typesofmaterials WHERE  typeofmaterial = '{self.poiskk.get()}' """
+            if a==2:
+                query=f"""SELECT * FROM materials WHERE  size  = '{self.poiskk.get()}' or characteristic= '{self.poiskk.get()}'"""
         db_rows = self.run_query(query,TRUE)
         for row in db_rows:
-            self.tree.insert('', 0, text = row[0], values = row[1])
-    
-    
+            if a==1:
+                self.tree.insert('', 0, text = row[0], values = row[1])
+            if a==2:
+                self.tree.insert('', 0, text=row[0], values=(row[1], row[2], row[3], row[4]))
+    def edit_records1(self, new_id, old_id, new_size, old_size, new_characteristic, old_characteristic, new_quantity, old_quantity, new_material_type, old_material_type):
+        
+        query = f"""UPDATE  materials SET id = {new_id}, size = '{new_size}', characteristic='{new_characteristic}',quantity ={new_quantity}, material_id ={new_material_type} WHERE id = {old_id}"""
+        self.run_query(query,False)
+        self.edit_wind.destroy()
+        self.message['text'] = 'мательял {} успешно изменен'.format(new_characteristic)
+        self.get_words()
 if __name__ == '__main__':
     window = Tk()
     application = Dictionary(window)
